Handle malformed URL properly
The old DetectCommonURLs checks if a string is a valid URL by validing given
string which has [0-9][a-z][A-Z][+-.] and "://".  That is incorrect because
the string "aaa://test.host" would be accepted as a valid one.

A helper function is added to perform further checking if the string has protocol
http://, https://, telnet://, and ftp:// or not.

Based on the original work of Wen Liao <zzz00072@gmail.com>