Add mouse browsing support
New mouse actions associated:
 - L-button to enter
 - M-button to leave
 - scroll to up and down

Since the functions are the same as original keyboard bindings,
instead of calling SendRawString(), new helper functions are
created for both keyboard and mouse.

In addition, most of the hardcoded SendRawString() calls in
view/telnetview.cpp are fixed.  The rest are single characters,
which can not be applied to SendString() call.

Signed-off-by: Jim Huang <jserv@0xlab.org>