Update sitelistdlg.cpp
Solve the problem of site list that the search can't reach the sites after first leaf directory.

https://github.com/pcman-bbs/pcmanx/issues/32

When user searches a site,
it would not incorrectly stop at the last site of the first leaf directory,
it will return to parent directory and continue the search.

Detail of the problem:
If "it" has no next,
"it" is set to be invalid by gtk_tree_model_iter_next( model, &it),
then gtk_tree_model_iter_next( model, &parent_it) would be wrong.