Shrink the title of an opened tab
If the title of an opened tab is too long, it will not be shown
within a tab. Instead, it shows all the characters of the title
and mess with the other tab title. Please see the following
picture for a better problem description.

http://i.imgur.com/MPPosvb.png

This commit first checks if the the length of a title is longer
than 20 characters. If so, it will shrink the title into 17
characters and add '...' at the end, 20 characters in total.

http://i.imgur.com/k8VggEN.png