Right-click popup menu only works on current tab
fix the issue #44, right-click on the tab and the popup menu may
behaves(close, reconnect, addtofavorite) on the wrong tab.

If users right-click on the tab, pcmanx would swtiches to the tab
first and pops up the menu.
It behaves like CMainFrame::OnCloseSelectCon (mouse middle click).

However, it wouldn't switch back to the original tab when user
selects "close" in the popup menu. Because we can't ensure which
one would be chosen.