Fix the incorrect site list after traversing first leaf directory
Fixes #32

When a user tries to search a BBS site, it was incorrect to stop
at the last site of the first leaf directory. Thus, it would return
to parent directory and continue the search.

Detail of the problem:
If the item has no next, it is set to be invalid by
gtk_tree_model_iter_next( model, &it), then
gtk_tree_model_iter_next( model, &parent_it) would be wrong.

Reported and fixed by b88923509